Internals

Complete docstring documentation of the JLPeg internal API.

JLpeg.Instruction — Type

Instruction

Abstract type of JLPeg VM instructions.

The possible struct fields have consistent meanings:

Field	Type	Meaning

op	@enum UInt8	The Opcode

l	`Int32	16`

n	UInt32	A subject index

c	Char	A character to match

b	UInt8	A test byte

tag	UInt32	Key in a capture/throw Dict

vec	BitVector (kinda)	A set's test BitVector

JLpeg.VMState — Type

VMState(patt::Pattern, string::AbstractString)

Contains the state of a match on subject by program.

Fields

Other than the subject and program, we have:

top: the byte length of subject

i: Instruction counter

s: Subject pointer

ti, ts, tc, tp: Stack top registers

t_on: flag for nonempty stack

inpred: Bool is true inside predicates (PAnd, PNot)

sfar: furthest subject pointer we've failed at, for error reporting

failtag: tag for a labeled failure

stack: Contains stack frames for calls and backtracks

cap: A stack for captures

running: Bool which is true when the VM is executing

matched: true if we matched the program on the subject

Implementation

Classic dispatch-driven VM with a program counter, opcodes, stack frames. I borrowed a page from Forth and made the top of the stack a register, allowing stack frames to be immutable while still permitting the PartialCommit optimization.