Reference

Reference

Core API

Pattern Matching

Constructing Patterns

JLpeg.Combinators

Captures and Actions

Rules and Grammars

Dialects

Generators

Core API

The basic structure of recognition in JLPeg is a Pattern.

JLpeg.Pattern — Type

Container for various patterns and grammars. Always has val, which may be primitive or a PVector, and code, an IVector. Some patterns have a field unique to that pattern type.

These are created in a wide variety of ways, all based on three functions, P, S, and R, with the rest of the work done with a great quantity of operator overloading.

JLpeg.PegMatch — Type

PegMatch <: AbstractMatch

A type representing a single match to a Pattern. Typically returned from the match function. PegFail is the atypical return value.

A PegMatch is equal (==) to a Vector if the captures are equal to the Vector.

Properties:

subject stores the string matched.

last is the index of the last character matched by patt. Due to the nature of PEGs, the match always begins at the first character, so there is little point in storing this information as a SubString (although the subject itself may be one).

captures contains any captures from matching patt to subject. This Vector can in principle contain anything, as captures may call functions, in which case the return value of that function becomes the capture. For more information, consult the JLPeg documentation, and the docstrings for C, Cg, Cc, A, and Anow.

offsets is a Vector{Int}, provided for compatibility with AbstractMatch. SubStrings contain their own offsets, so this is unnecessary for normal work, but we JIT it for match.offsets. It then consists of the indices at which the outer layer of captures may be found within the subject.

patt is the Pattern matched against the subject.

JLpeg.PegFail — Type

PegFail

Returned on a failure to match(patt:Pattern, subject::AbstractString). errpos is the position at which the pattern ultimately failed to match, label is info about the failure provided by T(::symbol), defaulting to :default if the pattern fails but not at a throw point, or if no throws are provided.

JLpeg.PegError — Type

PegError(msg) <: Exception

An error while constructing a Pattern.

Pattern Matching

Base.match — Function

match(patt::Pattern, subject::AbstractString)::Union{PegMatch, PegFail}

Match patt to subject, returning a PegMatch implementing the expected interface for its supertype AbstractMatch, or a PegFail with useful information about the failure.

JLpeg.compile! — Function

compile!(patt::Pattern)::Pattern

Compile a Pattern.

Translate the Pattern to Instruction codes, appending them to the code field and returning same. Performs various optimizations in the process.

In most cases, the return value of compile! is the same Pattern passed in. The exceptions are special cases of primitive types, but be sure to reassign the return value to the variable bound to the original for the general case.

compile!(patt::PGrammar)::Pattern

Compiles and prepares a Grammar.

Constructing Patterns

Patterns are built by combining the products of these constructors into more complex Patterns.

JLpeg.P — Function

P(p::Union{AbstractString,AbstractChar,Integer,Bool,Symbol})::Pattern

Create a Pattern.

If p is a String or Char, this matches that string or character.

If p is a positive Integer, it matches that many characters.

If p is true, the rules succeeds, if false, the rule fails.

If p is a Symbol, this represents a call to the rule with that name.

If p is a negative Integer, matches if that many characters remain, consumes no input.

If p is already a Pattern, it is simply returned.

Examples

julia> match(P("func"), "funci")
PegMatch(["func"])

julia> match(P(3), "three")
PegMatch(["thr"])

julia> match(P(true), "pass")
PegMatch([""])

julia> match(P(false), "fail")
PegFail("fail", 1)

julia> match(P('👍'), "👍")
PegMatch(["👍"])

JLpeg.@P_str — Macro

P"str"

Calls P(str) on the String, in close imitation of Lua's calling convention.

JLpeg.S — Method

S(s::AbstractString)

Create a Pattern matching any charcter in the string.

JLpeg.@S_str — Macro

S"str"

Calls S(str) on the String, in close imitation of Lua's calling convention.

JLpeg.R — Method

R(s::AbstractString)

Create a Pattern matching every character in the range from the first to the second character. s must be two codepoints long, and the first must be lower-valued than the second.

JLpeg.R — Method

R(a::AbstractChar, b::AbstractChar)

Match any character in the range a to b, inclusive.

JLpeg.@R_str — Macro

R"str"

Calls R(str) on the String, in close imitation of Lua's calling convention.

JLpeg.B — Function

B(p::Union{Pattern,AbstractString,AbstractChar,Integer})

Match patt behind the current subject index. patt must be of fixed length. The most useful B pattern is !B(1), which succeeds at the beginning of the string, followed by B(' ') to match the start of a line.

JLpeg.Combinators

JLpeg.Combinators — Module

JLpeg.Combinators

Contains all the combinator operators which shadow definitions in Base. The module redefines these symbols, providing a fallback to the Base, and are designed not to break existing code. The Julia compiler handles this kind of redirection well during inference, this technique is used by several packages which specialize operators for numerically-demanding tasks, without issue; we even transfer the Base docstrings.

Whether or not these overloads rise to the level of piracy is debatable. That said, we've walled them off, so that debate is not necessary.

Use

The @grammar and @rule macros are able to use these operators, so you may not need them. To import them, add this to your module:

import JLpeg.Combinators: *, -, %, |, ^, ~, !, >>, >:, inv

Captures and Actions

PEG patterns are recognizers, matching the longest prefix of the string which they're able. Captures allow for substrings to be captured within the match, Actions perform actions at the index of the match, or on a capture.

JLpeg.C — Function

C(patt::Pattern)

Create a capture. Matching patt with return the matched substring.

C(patt::Pattern, sym::Union{Symbol,AbstractString})

Create a named capture with key :sym or "sym".

JLpeg.Cg — Function

Cg(patt::Pattern [, sym::Union{Symbol,AbstractString}])

Create a group capture, which groups all captures from P into a vector inside the PegMatch object. If sym is provided, the group will be found at that key.

JLpeg.Cr — Function

Cr(patt::Pattern, sym::Union{CapSym, Nothing})

Captures a UnitRange of matches in patt, optionally keyed by sym. Convenient for substitutions and annotations.

JLpeg.Cc — Function

Cc(args...)

Constant capture. Matches the empty string and puts the values of args as a tuple in that place within the PegMatch captures. If args is a Pair{Symbol,Any}, that symbol will appear as a key in the match.

JLpeg.Cp — Function

Cp()

Captures the empty string in .captures, consuming no input. Useful for the side effect, of storing the corresponding offset.

JLpeg.A — Function

A(patt::Pattern, fn::Function)

Acts as a grouping capture for patt, applying fn to a successful match with the captures as arguments (not as a single Vector). If patt contains no captures, the capture is the SubString. The return value of fn becomes the capture; if nothing is returned, the capture (and its offset) are deleted.

JLpeg.Anow — Function

Anow(patt::Pattern, fn::Function)

Acts as a grouping capture, applying fn immediately upon a match succeeding to the captures inside patt, or the span of patt if there are no captures within it. The return value of fn becomes the capture, but if nothing is returned, the entire pattern fails.

JLpeg.T — Function

T(label::Symbol)

Throw a failure labeled with :label. If a rule :label exists, this will be called to attempt recovery, the success or failure of the recovery rule is then used. Otherwise, :label and the position of T(:label) will be attached to PegFail in the event that the whole match fails.

Rules and Grammars

The great advantage (other than composability) PEGs have over regular expressions is the ability to match recursive patterns. These are constructed out of Rules and Grammars.

JLpeg.Grammar — Type

Grammar(rule...)

Create a grammar from the provided rules.

JLpeg.Rule — Type

Rule(name::Symbol, patt::Pattern)

Un-sugared form of @rule, creates a rule from patt, assigning it the name name.

JLpeg.@rule — Macro

@rule :name ← pattern...

Sugared form for rule definition. Assigns the rule in-scope with the given name:

# Wrong
name = @rule :name  ←  "foo" | "bar"
# Right
@rule :name ← "foo" | "bar"

JLpeg.@grammar — Macro

@grammar(name, rules)

Syntax sugar for defining a set of rules as a single grammar. Expects a block rules, each of which is a rule-pair as can be created with ←, or, if you must, <=. "string" will be interpolated as P("string"), and :symbol as P(:symbol). Be sure to use the macro forms S"123" and R"az" for sets and ranges, which will otherwise be transformed into S(P("123")), which is invalid.

Example use

This simple grammar captures the first string of numbers it finds:

julia> @grammar capnums begin
            :nums  ←  (:num,) | 1 * :nums
            :num   ←  S"123"^1
        end;

julia> match(capnums, "abc123abc123")
PegMatch(["123"])

This one also captures the lowercase letters, before converting them to uppercase.

julia> upper = uppercase;  # A thoroughly unhygienic macro

julia> @grammar uppernums begin
           :nums  ←  (:num,) | :abc * :nums
           :num   ←  S"123"^1
           :abc   ←  R"az"^1 <| upper
       end;

julia> match(uppernums, "abc123abc123")
PegMatch(["ABC", "123"])

More extensive examples may be found in the documentation.

Dialects

A work in progress.

JLpeg.re — Constant

re : An Interpretation of LPeg's `re` module

The first dialect, intended, among other things, as a useful bootstrap of other dialects.

Generators

A PEG is a specification of a class of algorithms which are valid on a universe of strings. While the common thing to do is use this specification to construct a recognizer for such strings, it may also be used to create generators for valid strings in that universe.

JLpeg aspires to provide a complete set of generators for our patterns. So far, we have:

JLpeg.generate — Function

generate(set::PSet)::String

Generate a String of all characters matched by a Set.